# openapi3-v1.yaml
openapi: 3.0.1
info:
  title: Job Manager Service - API v1
  description: Job Manager Service API version 1
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  /v1/jobs/{jobId}/user-metadata:
    patch:
      operationId: updateUserMetadata
      parameters:
        - $ref: '#/components/parameters/jobId'
      summary: Update job's custom metadata
      description: >
        Updates the user-defined metadata object for a specific job.

        This endpoint allows clients to attach or modify arbitrary data related
        to a job

        without affecting the job's core properties or execution status.


        User metadata is useful for storing application-specific context,
        tracking information,

        or any custom data needed by client applications.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userMetadata'
      responses:
        '200':
          description: User metadata successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid metadata format or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
...
  /v1/jobs/{jobId}/status:
    parameters:
      - $ref: '#/components/parameters/jobId'
    put:
      operationId: updateStatus
      summary: Update job status for workflow control
      description: >
        Updates the operational status of a job to control its workflow
        execution. This endpoint 

        supports user-initiated operations including PENDING (resume), ABORTED
        (cancel), and 

        PAUSED (suspend), which may cascade changes to related stages and tasks.


        Internal status transitions (such as IN_PROGRESS, COMPLETED, FAILED,
        CREATED) are managed 

        automatically by the system based on stage and task completion, and
        cannot be set through 

        this endpoint.


        Status changes follow a state machine that enforces valid transitions.
        When a job's status 

        is changed, the system will automatically update timestamps and
        completion percentages as 

        appropriate.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/jobOperationStatus'
              required:
                - status
      responses:
        '200':
          description: Job status successfully changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid status or illegal state transition
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for bad request errors, providing
                      a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
                          - ILLEGAL_JOB_STATUS_TRANSITION
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/jobs/{jobId}/stages:
    parameters:
      - $ref: '#/components/parameters/jobId'
    get:
      operationId: getStagesByJobId
      parameters:
        - $ref: '#/components/parameters/includeTasks'
      summary: Retrieve all stages for a specific job in execution order
      description: >
        Fetches all stages associated with the specified job ID, automatically
        ordered by 

        their execution sequence (order field). Provides complete information
        about each 

        stage including type, status, progress, and sequential order number.


        Stages are returned in ascending order based on their order field (1, 2,
        3, etc.)

        to maintain the correct workflow execution sequence defined when stages
        were created.


        Optional inclusion of related task data via the should_return_tasks
        parameter,

        allowing clients to retrieve the complete job hierarchy in a single
        request.
      tags:
        - jobs
      responses:
        '200':
          description: Successfully retrieved stages for the specified job
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/getStageResponse'
        '400':
          description: Invalid job ID format or other parameter error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/jobs/{jobId}/stage:
    post:
      operationId: addStage
      summary: Add a new stage as the last stage in the job workflow
      description: >
        Appends a new stage to an existing job with automatic order assignment.

        The stage will be added after any existing stages in the job's workflow
        sequence.

        Each stage is automatically assigned an incremental order number (1, 2,
        3, etc.) 

        within the scope of its parent job to maintain execution sequence.


        This endpoint allows for extending job workflows at runtime by adding
        new processing steps.

        The order field ensures stages are processed in the correct sequence
        when retrieved.


        The job must exist and be in a valid state to accept new stages.
      parameters:
        - $ref: '#/components/parameters/jobId'
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createStagePayloadRequest'
            examples:
              processing_stage:
                summary: Data processing stage
                description: Stage for data transformation and validation
                value:
                  type: data_processing
                  data:
                    processing_type: batch
                    chunk_size: 1000
                    validation_rules:
                      - 'not_null: [id, timestamp]'
                      - 'range: [amount, 0, 1000000]'
                  userMetadata:
                    estimated_duration: 2h
                    resource_requirements:
                      cpu: 4 cores
                      memory: 8GB
              waiting_stage:
                summary: Manual approval stage
                description: Stage that waits for manual intervention
                value:
                  type: manual_approval
                  startAsWaiting: true
                  data:
                    approval_type: quality_check
                    reviewers:
                      - senior-analyst@company.com
                    checklist:
                      - Data quality metrics within acceptable range
                      - No PII detected in output
                      - Performance benchmarks met
                  userMetadata:
                    priority: high
                    sla: 24h
              stage_with_tracing:
                summary: Stage with distributed tracing
                description: Stage creation with provided tracing context
                value:
                  type: data_validation
                  data:
                    validation_type: schema_check
                    strict_mode: true
                  traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
                  tracestate: vendor=trace123,service=validation
      responses:
        '201':
          description: Stage successfully created and added to the job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stageResponse'
              examples:
                processing_stage_response:
                  summary: Response for data processing stage creation
                  value:
                    id: e9db7d16-a84c-4297-a45d-e59665760084
                    type: data_processing
                    status: CREATED
                    percentage: 0
                    jobId: 550e8400-e29b-41d4-a716-446655440000
                    order: 1
                    data:
                      processing_type: batch
                      chunk_size: 1000
                      validation_rules:
                        - 'not_null: [id, timestamp]'
                        - 'range: [amount, 0, 1000000]'
                    userMetadata:
                      estimated_duration: 2h
                      resource_requirements:
                        cpu: 4 cores
                        memory: 8GB
                    summary:
                      pending: 0
                      inProgress: 0
                      completed: 0
                      failed: 0
                      created: 0
                      retried: 0
                      total: 0
                    traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-33c489dd6db902b7-01
                    tracestate: processing=batch001,stage=data_proc
                waiting_stage_response:
                  summary: Response for manual approval stage creation
                  value:
                    id: 4f45bec6-08c4-4f7a-a3ea-be65eddc1235
                    type: manual_approval
                    status: WAITING
                    percentage: 0
                    jobId: 550e8400-e29b-41d4-a716-446655440000
                    order: 2
                    data:
                      approval_type: quality_check
                      reviewers:
                        - senior-analyst@company.com
                      checklist:
                        - Data quality metrics within acceptable range
                        - No PII detected in output
                        - Performance benchmarks met
                    userMetadata:
                      priority: high
                      sla: 24h
                    summary:
                      pending: 0
                      inProgress: 0
                      completed: 0
                      failed: 0
                      created: 0
                      retried: 0
                      total: 0
                    traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-44d59aee7ec902b7-01
        '400':
          description: Invalid request format.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for bad request errors, providing
                      a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
                          - JOB_IN_FINITE_STATE
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/stages:
    get:
      operationId: getStages
      parameters:
        - $ref: '#/components/parameters/paramJobId'
        - $ref: '#/components/parameters/paramStageType'
        - $ref: '#/components/parameters/stageStatus'
        - $ref: '#/components/parameters/includeTasks'
      summary: Retrieve stages matching specified criteria
      description: >
        Returns a filtered list of stages based on the provided query
        parameters.

        Supports filtering by job ID, stage type, and status.


        Optional inclusion of related task data via the should_return_tasks
        parameter

        allows clients to retrieve the complete stage hierarchy in a single
        request.
      tags:
        - stages
      responses:
        '200':
          description: Successfully retrieved matching stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/getStageResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/stages/{stageId}:
    parameters:
      - $ref: '#/components/parameters/stageId'
    get:
      operationId: getStageById
      parameters:
        - $ref: '#/components/parameters/includeTasks'
      summary: Retrieve a specific stage by its ID
      description: >
        Fetches detailed information about a stage using its unique identifier.

        Includes stage configuration, status, metadata, and completion
        information.


        Optional inclusion of related task data via the should_return_tasks
        parameter,

        which allows clients to retrieve the complete stage hierarchy in a
        single request.
      tags:
        - stages
      responses:
        '200':
          description: Stage data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getStageResponse'
        '400':
          description: Invalid request format or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stageNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/stages/{stageId}/summary:
    get:
      operationId: getStageSummary
      parameters:
        - $ref: '#/components/parameters/stageId'
      summary: Get stage progress summary
      description: >
        Retrieves aggregated statistics about the tasks within a specific stage.

        Provides counts of tasks by status (pending, in progress, completed,
        etc.) 

        and a total task count for monitoring stage progress.


        This endpoint is useful for displaying progress indicators or status
        dashboards

        without needing to retrieve and process all individual task details.
      tags:
        - stages
      responses:
        '200':
          description: Stage summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/summary'
        '400':
          description: Invalid stage ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stageNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/stages/{stageId}/user-metadata:
    patch:
      operationId: updateStageUserMetadata
      parameters:
        - $ref: '#/components/parameters/stageId'
      summary: Update stage's custom metadata
      description: >
        Updates the user-defined metadata object for a specific stage.

        This endpoint allows clients to attach or modify arbitrary data related
        to a stage

        without affecting the stage's core properties or execution status.


        User metadata is useful for storing application-specific context,
        tracking information,

        or any custom data needed by client applications.
      tags:
        - stages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userMetadata'
      responses:
        '200':
          description: User metadata successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: STAGE_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid metadata format or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stageNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/stages/{stageId}/status:
    put:
      operationId: updateStageStatus
      parameters:
        - $ref: '#/components/parameters/stageId'
      summary: Update stage status to trigger workflow progression
      description: >
        Updates the operational status of a stage to PENDING, which may trigger
        workflow 

        transitions and cascade changes to related tasks and the parent job.
        This endpoint 

        is restricted to manually advancing stages in the workflow sequence.


        Common use cases:

        - Resuming a WAITING stage (e.g., after manual intervention or approval)

        - Unblocking workflow progression when a stage is ready to proceed


        Internal status transitions (such as IN_PROGRESS, COMPLETED, FAILED,
        ABORTED, 

        WAITING, CREATED) are managed automatically by the system based on task
        completion, 

        job state, and workflow rules.


        Status changes follow a state machine that enforces valid transitions.
        When a stage's 

        status changes, it may automatically update the parent job's status and
        trigger 

        transitions in subsequent stages (e.g., activating the next stage in the
        sequence).
      tags:
        - stages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/stageOperationStatus'
              required:
                - status
      responses:
        '200':
          description: Stage status successfully changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: STAGE_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid status or illegal state transition
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for bad request errors, providing
                      a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
                          - ILLEGAL_STAGE_STATUS_TRANSITION
        '404':
          description: Stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stageNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for internal server errors,
                      providing a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - DATABASE_RELATED_ERROR
                          - UNKNOWN_ERROR
                          - ILLEGAL_JOB_STATUS_TRANSITION
                          - JOB_NOT_FOUND
...
security: []
components:
  parameters:
    jobId:
      in: path
      name: jobId
      description: Unique identifier for the job
      required: true
      schema:
        $ref: '#/components/schemas/jobId'
    stageId:
      in: path
      name: stageId
      description: Unique identifier for the stage
      required: true
      schema:
        $ref: '#/components/schemas/stageId'
...
    includeTasks:
      in: query
      name: should_return_tasks
      description: When true, includes task data in the response
      required: false
      schema:
        $ref: '#/components/schemas/returnTask'
...
    paramJobId:
      in: query
      name: job_id
      description: Filter results by job identifier
      required: false
      schema:
        $ref: '#/components/schemas/jobId'
    paramStageType:
      in: query
      name: stage_type
      description: Filter results by stage type (e.g., processing, validation)
      required: false
      schema:
        $ref: '#/components/schemas/stageType'
...
    stageStatus:
      in: query
      name: stage_operation_status
      description: |
        Filter results by stage operational status (e.g., PENDING, IN_PROGRESS).
        Used to find stages in specific execution states.
      required: false
      schema:
        $ref: '#/components/schemas/stageOperationStatusResponse'
  schemas:
...
    jobId:
      type: string
      format: uuid
      description: Unique identifier for a job
...
    stageId:
      type: string
      format: uuid
      description: Unique identifier for a stage
...
    successMessages:
      type: string
      enum:
        - JOB_MODIFIED_SUCCESSFULLY
        - TASK_MODIFIED_SUCCESSFULLY
        - STAGE_MODIFIED_SUCCESSFULLY
        - JOB_DELETED_SUCCESSFULLY
      example: JOB_MODIFIED_SUCCESSFULLY
      description: Success response codes
    jobOperationStatus:
      type: string
      enum:
        - PENDING
        - ABORTED
        - PAUSED
      example: PENDING
      description: >
        User-controllable job statuses: PENDING (resume), ABORTED (cancel),
        PAUSED (suspend).

        System-managed: IN_PROGRESS, COMPLETED, FAILED, CREATED.
...
    stageOperationStatus:
      type: string
      enum:
        - PENDING
      description: >
        User-controllable stage status: PENDING only.

        System-managed: IN_PROGRESS, COMPLETED, FAILED, ABORTED, WAITING,
        CREATED.
...
    userMetadata:
      type: object
      additionalProperties: true
      default: {}
      description: >-
        Application-specific custom data container that can store arbitrary
        client information without affecting core operations
...
    createStagePayloadRequest:
      allOf:
        - $ref: '#/components/schemas/createStagePayload'
        - type: object
          properties:
            startAsWaiting:
              type: boolean
              description: >
                Optional flag indicating whether the stage should be created in
                a waiting state.

                If true, the stage will not start processing immediately and
                will require

                manual intervention to begin execution. Useful for staging
                workflows where

                stages need to be prepared but not executed until all
                dependencies are met.
              example: false
          description: >
            Input payload for creating a new processing stage.

            Defines the stage type, custom configuration data, and user-defined
            metadata.

            Used when adding stages to jobs or creating stages as part of job
            creation.

            The system automatically assigns an incremental order number to
            maintain execution sequence.

            The startAsWaiting flag allows for creating stages that are not
            immediately active,

            enabling more complex workflows where stages can be prepared in
            advance.
...
    stageResponse:
      allOf:
        - $ref: '#/components/schemas/createStagePayload'
        - type: object
          required:
            - id
            - summary
            - jobId
            - order
            - traceparent
            - userMetadata
            - percentage
            - status
          properties:
            id:
              $ref: '#/components/schemas/stageId'
            summary:
              $ref: '#/components/schemas/summary'
            percentage:
              $ref: '#/components/schemas/percentage'
            status:
              $ref: '#/components/schemas/stageOperationStatusResponse'
            jobId:
              $ref: '#/components/schemas/jobId'
            order:
              $ref: '#/components/schemas/order'
            traceparent:
              $ref: '#/components/schemas/traceparent'
          description: >
            Complete stage information returned by the API, including all fields
            from the 

            creation payload plus system-generated fields like ID, progress
            statistics,

            execution status, and parent job reference.
    getStageResponse:
      allOf:
        - $ref: '#/components/schemas/stageResponse'
        - type: object
          description: >
            Extended stage response that includes associated tasks when
            requested with 

            should_return_tasks=true. Used to retrieve the complete stage
            hierarchy

            in a single API call.
          properties:
            tasks:
              type: array
              description: Associated tasks belonging to this stage
              items:
                $ref: '#/components/schemas/taskResponse'
...
    defaultOkMessage:
      type: object
      description: |
        Standard success response structure used for operations that don't
        return entity data, providing a standardized confirmation message.
      properties:
        code:
          $ref: '#/components/schemas/successMessages'
      required:
        - code
    baseErrorResponse:
      type: object
      description: >
        Foundation schema for standardized error responses across all API
        endpoints.

        Provides consistent structure with human-readable message and
        common-readable code fields

        that are extended by individual endpoints to define their specific error
        scenarios.
      properties:
        message:
          type: string
          description: Human-readable error message describing the issue
        code:
          type: string
          description: common-readable error code indicating the specific error type
      required:
        - message
        - code
    internalErrorsResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Provides a unified error response format for internal server errors,
            including a common-readable error code and a clear, human-readable
            message to help clients identify and handle issues consistently.
          properties:
            code:
              enum:
                - DATABASE_RELATED_ERROR
                - UNKNOWN_ERROR
          required:
            - code
    badRequestResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response structure for bad request errors,
            providing a consistent

            format for common-readable error codes and human-readable messages.
          properties:
            code:
              enum:
                - VALIDATION_ERROR
          required:
            - code
    jobNotFoundResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response for cases where a requested job is not
            found in the system.

            Provides a consistent format for common-readable error codes and
            human-readable messages.
          properties:
            code:
              enum:
                - JOB_NOT_FOUND
          required:
            - code
    stageNotFoundResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response for cases where a requested stage is not
            found in the system.

            Provides a consistent format for common-readable error codes and
            human-readable messages.
          properties:
            code:
              enum:
                - STAGE_NOT_FOUND
          required:
            - code
...